Dimension hierarchy as a data frame

Dimension hierarchy as a data frame

getHierarchy(stage, parent_id = NA, nstage = 0)

Arguments

stage
a hierarchical list with dimension stage and it's children.

parent_id
Id of the parent of the stage.

nstage
The depth of the stage relative to the root (which is 0).

Recursively retrieve all children of the dimension and flatten as a single data frame. Helper for parse_dimensions()

Note

Kiitos HY TIRA-kurssi 2013

Author

Tuomo Nieminen

Examples

url <- "https://sampo.thl.fi/pivot/prod/en/epirapo/covid19case/fact_epirapo_covid19case.json"
dimensions <- openthl:::get_dimensions(url)

df <- openthl:::getHierarchy(dimensions$children[[1]], parent_id = dimensions$id[[1]])
str(df)

#> 'data.frame':	310 obs. of  15 variables:
#>  $ stage1_id       : chr  "hcdmunicipality20202" "hcdmunicipality20202" "hcdmunicipality20202" "hcdmunicipality20202" ...
#>  $ stage1_sid      : int  445131 445131 445131 445131 445131 445131 445131 445131 445131 445131 ...
#>  $ stage1_label    : chr  "Åland" "Åland" "Åland" "Åland" ...
#>  $ stage1_stage    : chr  "hcd" "hcd" "hcd" "hcd" ...
#>  $ stage1_code     : chr  "hcdmunicipality20202" "hcdmunicipality20202" "hcdmunicipality20202" "hcdmunicipality20202" ...
#>  $ stage1_sort     : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ stage1_uri      : chr  "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/hcd/2" "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/hcd/2" "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/hcd/2" "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/hcd/2" ...
#>  $ stage1_parent_id: chr  "hcdmunicipality20201" "hcdmunicipality20201" "hcdmunicipality20201" "hcdmunicipality20201" ...
#>  $ stage2_id       : chr  "hcdmunicipality202023" "hcdmunicipality202024" "hcdmunicipality202025" "hcdmunicipality202026" ...
#>  $ stage2_sid      : int  445268 444988 445090 445231 445227 445124 445080 445266 445153 445091 ...
#>  $ stage2_label    : chr  "Brändö" "Eckerö" "Finström" "Föglö" ...
#>  $ stage2_stage    : chr  "municipality" "municipality" "municipality" "municipality" ...
#>  $ stage2_code     : chr  "hcdmunicipality202023" "hcdmunicipality202024" "hcdmunicipality202025" "hcdmunicipality202026" ...
#>  $ stage2_sort     : int  23 24 25 26 27 28 29 30 31 32 ...
#>  $ stage2_uri      : chr  "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/municipality/23" "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/municipality/24" "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/municipality/25" "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/municipality/26" ...
#>  - attr(*, "root")='data.frame':	1 obs. of  8 variables:
#>   ..$ id       : chr "hcdmunicipality20201"
#>   ..$ sid      : int 445222
#>   ..$ label    : chr "All areas"
#>   ..$ stage    : chr "root"
#>   ..$ code     : chr "hcdmunicipality20201"
#>   ..$ sort     : int 1
#>   ..$ uri      : chr "http://meta.thl.fi/codes/wild/dimension/hcdmunicipality2020/1/root/1"
#>   ..$ parent_id: chr "hcdmunicipality2020"
#>  - attr(*, "nstage")= num 2